four nested loops

+++ 0x01-variables_if_else_while/102-print_comb5.c
@@ -25,7 +25,7 @@ int main(void)
 			{
 				for (m = '0'; m <= '9'; m++)
 				{
-					if ((i != n || j != m) && ((int)i * 10 + (int)j < (int)n * 10 + (int)m))
+					if ((i != n || j != m) && ((i * 10 + j) < (n * 10 + m)))
 					{
 					putchar(i);
 					putchar(j);
